Fix Carbon addHours type error - cast config to int
The config() helper returns strings from .env files, but Carbon's
addHours() expects int|float. Cast the retention hours to int to
prevent TypeError.

+++ app/Http/Controllers/Api/MeetingController.php
@@ -43,7 +43,7 @@ class MeetingController extends Controller
             'user_id' => auth()->id(),
             'state' => 'created',
             'retention_until' => auth()->guest()
-                ? now()->addHours(config('app.demo_retention_hours', 48))
+                ? now()->addHours((int) config('app.demo_retention_hours', 48))
                 : null,
         ]));
